backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pymupdf==1.23.8
pymupdf4llm==0.0.17
sentence-transformers==2.2.2
faiss-cpu==1.7.4
sqlalchemy==2.0.23
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
openpyxl==3.1.2
